Refactor inner class to standalone class
