+++ src/Matcher/DeferredError.php
@@ -1,0 +1,39 @@
+<?php
+/*
+ * This file is part of the phpunit-mock-objects package.
+ *
+ * (c) Sebastian Bergmann <sebastian@phpunit.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+namespace PHPUnit\Framework\MockObject\Matcher;
+use PHPUnit\Framework\MockObject\Invocation as BaseInvocation;
+
+class DeferredError extends StatelessInvocation
+{
+    /**
+     * @var \Throwable
+     */
+    private $exception;
+
+    public function __construct(\Throwable $exception)
+    {
+        $this->exception = $exception;
+    }
+
+    public function verify()
+    {
+        throw $this->exception;
+    }
+
+    public function toString(): string
+    {
+        return '';
+    }
+
+    public function matches(BaseInvocation $invocation)
+    {
+        return true;
+    }
+}
